refactor: restructure preferences services and screen (#49)
refactor: restructure preferences services and screen
Introduced a new `PreferencesService` class for managing application settings.

fix: add `enableAdvancedSettings` setting check for initialization tasks
Added a check for the `enableAdvancedSettings` setting to ensure initialization tasks are performed correctly.

fix: update keyboard layout only if no related advanced settings is on
Fixed an issue where the keyboard layout would still update if related advanced settings are enabled.

fix: show alt layout on startup only if advanced settings is on
Ensured that the alternate keyboard layout is not displayed on startup when advanced settings are not enabled.

fix:  correct user layout loading with Kanata and advanced settings
Fixed an issue where the user layout would incorrectly load when Kanata is enabled and advanced settings are disabled.

fix: trigger hide timer on app 
Added a delay to start the auto-hide timer when the application is opened and auto-hide is enabled.